🛠️ pipeline.py -> Removed unused IconDirectoryMap import 🔴 src/iconmap.py -> Deleted IconDirectoryMap class file 🛠️ src/iconmatch.py -> Updated SSIMEngine import path 🛠️ src/matching/ssim/__init__.py -> Renamed SSIM engine module 🛠️ src/matching/ssim/common.py -> Renamed common file and updated imports 🛠️ src/matching/ssim/match.py -> Renamed match file and updated imports 🛠️ src/matching/ssim/quality.py -> Renamed quality file and updated imports
